% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfix-calculator.R
\name{calculate_postfix}
\alias{calculate_postfix}
\alias{check_valid_operators}
\title{This function calculates postfix expression}
\usage{
check_valid_operators(operator)
}
\arguments{
\item{input_string}{A string with space-delimited integer arithmetic expression in postfix notation.}

\item{char}{This function checkd if the character value is \code{+, -, *, /.}}

\item{check_number}{chcks if input is a number}

\item{x}{operand 1}

\item{y}{operator}

\item{z}{operand 2}

\item{input_vector}{input vector}

\item{values}{numeric values}
}
\value{
A integer value
}
\description{
This program computes the value by accepting the string input which is space-delimited
             arithmetic expression in postfix notation.
}
\examples{

calculate_postfix("2 3 1 * + 9 - ")
}
\author{
Som Bohora
}
