% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize-each-word.R
\name{capitalize_each_word}
\alias{capitalize_each_word}
\title{This function makes first letter of each word in a string to uppercase}
\usage{
capitalize_each_word(x)
}
\arguments{
\item{x}{A string object or a vector of string objects with or without . or _ or space
delimiters}
}
\value{
A string with first letter of each word capitalized
}
\description{
This function capitalizes the first letter of each word in a string to uppercase
}
\examples{

capitalize_each_word("make each letter in this string capitalized.")
capitalize_each_word("make_each _etter_in_this_string_capitalized.")
capitalize_each_word("make.each.letter.in.this.string.cpitalized.")
}
\author{
Som Bohora
}
