% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci-recursion.R
\name{compute_fibonacci}
\alias{compute_fibonacci}
\title{This function returns the Fibonacci number at a given position}
\usage{
compute_fibonacci(n)
}
\arguments{
\item{n}{A position for which the Fibonacci number is desired}
}
\value{
A Fibonacci number at a given position
}
\description{
This function returns the Fibonacci number at a given position using recursion method
}
\examples{
compute_fibonacci(5)
n = 10
for (i in 0:(n-1)) {
  print(compute_fibonacci(i))
}
}
\author{
Som Bohora
}
