% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-words-in-a-string.R
\name{count_words_in_string}
\alias{count_words_in_string}
\title{This function count the number of words in given string}
\usage{
count_words_in_string(string, pattern = "[.]", remove_punctuation = FALSE)
}
\arguments{
\item{string}{A text or paragraph}

\item{pattern}{A pattern to split the given string with}

\item{remove_punctuation}{A boolean argument whether to remove punctuation characters}
}
\value{
Numeric vector of words with their counts of occurrence in passed string
}
\description{
This function count the number of words in given string by removing or without special
punctuation characters.
}
\examples{

string = "This is a string. This is a a word. hora?"
count_words_in_string(string = string, pattern = "[ ]", remove_punctuation = TRUE)
}
\author{
Som Bohora
}
