% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci-loop.R
\name{fibonacci_loop}
\alias{fibonacci_loop}
\title{This function returns the Fibonacci number at a given position}
\usage{
fibonacci_loop(n)
}
\arguments{
\item{n}{A position for which the Fibonacci number is desired}
}
\value{
A Fibonacci number at a given position
}
\description{
This function returns the Fibonacci number at a given position using loop method
}
\examples{

fibonacci_loop(3)
}
\author{
Som Bohora
}
