% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-max.R
\name{find_max}
\alias{find_max}
\title{This function finds the maximum value from a vector}
\usage{
find_max(vector)
}
\arguments{
\item{vector}{A vector of values}
}
\value{
A maximum value from the given vector
}
\description{
This function finds the maximum value from a vector
}
\examples{

find_max(c(2,3,200,5,6,7))
}
\author{
Som Bohora
}
