% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-mode.R
\name{find_mode}
\alias{find_mode}
\title{This function finds the mode value from a vector}
\usage{
find_mode(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A mode value from the given vector
}
\description{
This function finds the mode value from a vector
}
\examples{

find_mode(c(2,2,3,2,5,5,5,5,6,7))
}
\author{
Som Bohora
}
