% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfix-calculator.R, R/prefix-infix.R
\name{calculate_postfix}
\alias{calculate_postfix}
\alias{check_valid_operators}
\alias{check_operator}
\alias{check_digit}
\alias{perform_operation}
\alias{push}
\alias{pop}
\alias{calculate_postfix}
\alias{evaluate_prefix}
\alias{check_digit}
\alias{evaluate_prefix}
\title{This function calculates postfix expression}
\usage{
check_valid_operators(operator)

check_operator(char)

check_digit(number)

perform_operation(z, x, y)

push(input_vector, values)

pop(input_vector)

calculate_postfix(input_string)

check_digit(number)

evaluate_prefix(expression)
}
\arguments{
\item{operator}{A string with mathematical operators}

\item{char}{This function checkd if the character value is}

\item{number}{checks if input is a number}

\item{z}{operand 2}

\item{x}{operand 1}

\item{y}{operator}

\item{input_vector}{input vector}

\item{values}{numeric values}

\item{input_string}{A string with space-delimited integer arithmetic expression in postfix notation.}

\item{expression}{A string with space-delimited integer arithmetic expression in prefix notation}
}
\value{
A integer value

A integer value
}
\description{
This program computes the value by accepting the string input which is space-delimited
             arithmetic expression in postfix notation.

This program computes the value by accepting the string input which is space-delimited
             arithmetic expression in prefix notation.
}
\examples{

calculate_postfix("2 3 1 * + 9 - ")

expr <- "+ 9 * - 2 5 + 3 2"
evaluate_prefix(expr)
}
\author{
Som Bohora

Som Bohora
}
