% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertion-sort.R
\name{insertion_sort}
\alias{insertion_sort}
\title{This function sorts the unsorted vector}
\usage{
insertion_sort(input_vector)
}
\arguments{
\item{input_vector}{An unsorted vector}
}
\value{
A sorted vector
}
\description{
This function sorts the given unsorted vector using insertion algorithm
}
\examples{

x <- c(2,1,4,7,5,4,2,1,3,1,9,8,7)
insertion_sort(x)
}
\author{
Som Bohora
}
