% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palindrome.R
\name{is_palindrome}
\alias{is_palindrome}
\title{This function checks if the given string is a palindrome}
\usage{
is_palindrome(str)
}
\arguments{
\item{str}{A string or full sentence}
}
\value{
TRUE if s is a palindrome and FALSE otherwise
}
\description{
Tests whether a string or sentence is a palindrome by removing special punctuation characters. A palindrome
is equal to its reverse, for example "rotor" or "racecar".
}
\examples{

is_palindrome("Was It A Rat I Saw?")
}
\author{
Som Bohora
}
