% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-pangram.R
\name{is_pangram}
\alias{is_pangram}
\title{This function checks if the given string is a pangram}
\usage{
is_pangram(str)
}
\arguments{
\item{str}{A string or full sentence}
}
\value{
TRUE if s is a pangram and FALSE otherwise
}
\description{
Tests whether a string or sentence is a pangram by removing special punctuation characters. A pangram
is a sentence constructed using all the letters of the alphabet.
}
\examples{

string = "The quick brown fox jumps over the lazy dog"
x = "Back in June we delivered oxygen equipement of the same size"

is_pangram(string)
is_pangram(x)
}
\author{
Som Bohora
}
