% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper-case.R
\name{make_proper_case}
\alias{make_proper_case}
\title{This function removes variable separatedwith or without space or underscore or dot delimiters 
and converts the string to proper case}
\usage{
make_proper_case(string)
}
\arguments{
\item{string}{A string with or without space or underscore or dot delimiters}
}
\value{
A string with proper case
}
\description{
This function removes variable separatedwith or without space or underscore or dot delimiters 
and converts the string to proper case
}
\examples{

make_proper_case("this is proper case.")
make_proper_case("this_is_proper_case.")
make_proper_case("this.is.proper.case.")
}
\author{
Som Bohora
}
