% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepalese-words-to-number.R
\name{nepalese_words_to_number}
\alias{nepalese_words_to_number}
\title{This function allows you to translate to integer numerical words spelled in Nepali.}
\usage{
nepalese_words_to_number(x)
}
\arguments{
\item{x}{x is an input string}
}
\value{
An integer value
}
\description{
This function allows you to translate to integer numerical words spelled in Nepali.
Text must be previously cleaned & removed extraneous words or symbols. Below is the list of nepalese
words used in this function;

 \code{"aath", "arba", "athara",
"athchaalees", "aththais", "athtis", "baais", "barha", "battis", "bayalees", "bis", "chalees", "char",
"chaubis", "chaudha", "chauntis", "chawaalees", "chha", "chhabbis", "chhattis", "chhayaalees", "das", 
"dui", "eghaara", "ek", "ekchalees", "ekkais", "ekkattis", "hajaar", "karod", "kharaba", "lakh", "nau", 
"nil", "paanch", "pachas", "pachchis", "padhna", "paintis", "paitaalees", "pandhra", "sacchaalees", 
"saitis", "sat", "satra", "sattais", "saya", "shankha", "soonya", "sorha", "teis", "terha", "tettis",
"tin", "tis", "triyalees", "unanchas", "unantis", "unnais", "unnanchaalis"}
}
\examples{

x = "ek karod paanch hajaar paanch saya paanch"  # 10005505
nepalese_words_to_number(x)
}
\author{
Som Bohora
}
