% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-group-means.R
\name{plot_group_means}
\alias{plot_group_means}
\title{Calculate averages by grouping variables and plot means}
\usage{
plot_group_means(data, group1, group2 = NULL, group3 = NULL, y)
}
\arguments{
\item{data}{A datadrame (not string)}

\item{group1}{First grouping variable (string)}

\item{group2}{Second grouping variable (string)}

\item{group3}{Third grouping variable (string)}

\item{y}{numeric variable for which grouped is to be calculated}
}
\value{
A plot of averages by grouping variable/s
}
\description{
Calculate averages by one, two or three grouping variables and plot means with confidence intervals aound means
}
\examples{
plot_group_means(data = ggplot2::diamonds, group1 = "color", group2 = "cut", y = "price")
}
\author{
Som Bohora
}
