% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-sort.R
\name{selection_sort}
\alias{selection_sort}
\title{This function sorts the unsorted vector}
\usage{
selection_sort(input_vector)
}
\arguments{
\item{input_vector}{An unsorted vector}
}
\value{
A sorted vector
}
\description{
This function sorts the given unsorted vector using insertion algorithm
}
\examples{

x <- c(6,3,2,5,1,0,7,10,5)
selection_sort(x)
}
\author{
Som Bohora
}
