% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-rows-to-columns.R
\name{split_rows_to_columns}
\alias{split_rows_to_columns}
\title{This function takes a vector and truns it into a data frame with multiple columns.}
\usage{
split_rows_to_columns(vector, number_rows, col_name_prefix = "name")
}
\arguments{
\item{vector}{A vector}

\item{number_rows}{Number of rows to split the vector into data frame}

\item{col_name_prefix}{A prefix for column names for the data frame}
}
\value{
A data frame with multiple columns
}
\description{
This function takes a vector and truns it into a data frame with multiple columns.
}
\examples{

split_rows_to_columns(mtcars$mpg, number_rows = 7, col_name_prefix = "ID List")
split_rows_to_columns(mtcars$mpg, number_rows = 1, col_name_prefix = "ID List")
}
\author{
Som Bohora
}
