% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-nonmissing.R
\name{sum_nonmissing}
\alias{sum_nonmissing}
\title{Calculate sum of non missing values rowwise.}
\usage{
sum_nonmissing(...)
}
\arguments{
\item{...}{A list of column names from a dataframe for which non-missings are to be calculated rowwise}
}
\value{
A an integer value of sum of non-missing values
}
\description{
Calculate sum of all non missing values rowwise within mutate function in dplyr package.
}
\examples{
library(magrittr)
iris[c(1,2,4), c(1,3,4)] <- NA
iris \%>\%
  dplyr::rowwise() \%>\%  # Notice the use of rowwise here
  dplyr::mutate(Sum.Non.Missings = sum_nonmissing(
  Sepal.Length, Sepal.Width, Petal.Length, Petal.Width))
}
\author{
Som Bohora
}
