% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap-case.R
\name{swap_case}
\alias{swap_case}
\title{This function converts all lowercase letters to uppercase letters and vice versa.}
\usage{
swap_case(string)
}
\arguments{
\item{string}{A text or paragraph}
}
\value{
A string with swapped cases
}
\description{
Given a string, convert all lowercase letters to uppercase letters and vice versa.
}
\examples{

s <- "thTs Is a 'R Case Swap Function 2'."
swap_case(s)
}
\author{
Som Bohora
}
